+++ python_boilerplate/configuration/application_configuration.py
@@ -1,11 +1,11 @@
 from loguru import logger
 from pyhocon import ConfigFactory, ConfigTree
 
-from python_boilerplate.function_collection import get_resources_dir
+from python_boilerplate.common.common_function import get_resources_dir
 
 # `application_conf` contains the configuration for the application.
 application_conf: ConfigTree = ConfigFactory.parse_file(
-    f"{get_resources_dir()}/application.conf"
+    get_resources_dir() / "application.conf"
 )