+++ tests/demo/test_pandas_usage.py
@@ -1,22 +1,32 @@
+from pathlib import Path
 from typing import Hashable, Iterable
 
 import numpy as np
-import pandas as pd
 from loguru import logger
-from pandas import DataFrame, DatetimeIndex, Series
+from pandas import DatetimeIndex, Series
 
-from python_boilerplate.function_collection import get_resources_dir
+from python_boilerplate.demo.pandas_usage import (
+    look_for_sony_published_games,
+    pandas_data_structure_date_range,
+    pandas_data_structure_series,
+    sony_published_video_games_path,
+    video_games,
+)
+
+# https://pandas.pydata.org/pandas-docs/stable/user_guide/10min.html#minutes-to-pandas
 
 
 def test_pandas_data_structure_series() -> None:
-    series: Series = pd.Series([1, 3, 5, np.nan, 6, 8])
+    series: Series = pandas_data_structure_series()
     assert series.dtype == np.dtype("float64")
+    assert len(series) == 6
     logger.info(f"series:\n{series}")
 
 
 def test_pandas_data_structure_date_range() -> None:
-    dates: DatetimeIndex = pd.date_range("2022-01-01", periods=6)
+    dates: DatetimeIndex = pandas_data_structure_date_range()
     assert dates.dtype == np.dtype("datetime64[ns]")
+    assert len(dates) == 6
     date1 = dates.array[0].date()
     logger.info(f"date 1: {date1}")
     logger.info(f"dates:\n{dates}")
@@ -26,9 +36,8 @@ def test_pandas_reading_csv() -> None:
     """
     Read a csv file.
 
-    See the `CSV & text files <https://pandas.pydata.org/pandas-docs/stable/user_guide/io.html#io-read-csv-table>`
+    See the CSV & text files https://pandas.pydata.org/pandas-docs/stable/user_guide/io.html#io-read-csv-table
     """
-    video_games: DataFrame = pd.read_csv(f"{get_resources_dir()}/video_games.csv")
     logger.info(f"video_games:\n{video_games}")
     logger.info(f"Head of video_games:\n{video_games.head()}")
     logger.info(f"Tail of video_games:\n{video_games.tail()}")
@@ -41,4 +50,12 @@ def test_pandas_reading_csv() -> None:
         logger.info(f"index: {index}, title: {row['Title']}")
         if index == 5:
             break
+    assert video_games is not None
     assert video_games.iloc[1]["Release.Console"] == "Sony PSP"
+
+
+def test_look_for_sony_published_games():
+    sony_published_games = look_for_sony_published_games()
+    assert sony_published_games is not None
+    assert len(sony_published_games) == 60
+    assert Path(sony_published_video_games_path).exists(), "CSV file NOT exists!"